Extract map query runner partial
